Clean-up code using RatingScore

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/task/meta/PlanMissionMeta.java
@@ -52,7 +52,7 @@ public class PlanMissionMeta extends MetaTask implements SettlementMetaTask {
     /** Task name */
     private static final String NAME = Msg.getString("Task.description.planMission"); //$NON-NLS-1$
 
-    private static final int START_FACTOR = 50;
+    private static final double START_FACTOR = 50;
   
 
     public PlanMissionMeta() {
